CCMSG-1447: Added integration tests for routing SMTs.

+++ src/test/java/io/confluent/connect/elasticsearch/integration/ElasticsearchConnectorBaseIT.java
@@ -15,6 +15,14 @@
 
 package io.confluent.connect.elasticsearch.integration;
 
+import java.io.IOException;
+import java.net.ConnectException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.apache.kafka.connect.json.JsonConverter;
 import org.apache.kafka.connect.storage.StringConverter;
 import org.apache.kafka.test.TestUtils;
@@ -28,26 +36,19 @@ import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.net.ConnectException;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 import io.confluent.connect.elasticsearch.ElasticsearchSinkConnector;
 import io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig;
 import io.confluent.connect.elasticsearch.helper.ElasticsearchContainer;
 import io.confluent.connect.elasticsearch.helper.ElasticsearchHelperClient;
 
+
 import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.CONNECTION_PASSWORD_CONFIG;
 import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.CONNECTION_URL_CONFIG;
 import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.CONNECTION_USERNAME_CONFIG;
-import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.IGNORE_KEY_CONFIG;
-import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.IGNORE_SCHEMA_CONFIG;
 import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.DATA_STREAM_DATASET_CONFIG;
 import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.DATA_STREAM_TYPE_CONFIG;
+import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.IGNORE_KEY_CONFIG;
+import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConfig.IGNORE_SCHEMA_CONFIG;
 import static org.apache.kafka.connect.json.JsonConverterConfig.SCHEMAS_ENABLE_CONFIG;
 import static org.apache.kafka.connect.runtime.ConnectorConfig.CONNECTOR_CLASS_CONFIG;
 import static org.apache.kafka.connect.runtime.ConnectorConfig.KEY_CONVERTER_CLASS_CONFIG;
@@ -81,6 +82,7 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   protected ElasticsearchHelperClient helperClient;
   protected Map<String, String> props;
   protected String index;
+  protected boolean needToCleanup = true;
 
   @AfterClass
   public static void cleanupAfterAll() {
@@ -103,7 +105,7 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   public void cleanup() throws IOException {
     stopConnect();
 
-    if (container.isRunning()) {
+    if (container.isRunning() && needToCleanup) {
       if (helperClient != null) {
         try {
           helperClient.deleteIndex(index, isDataStream);
@@ -144,6 +146,17 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
     writeRecords(NUM_RECORDS);
 
     verifySearchResults(NUM_RECORDS);
+  }
+
+  protected void assertConnectorFails(Map<String, String> props, String trace) throws Exception {
+    // start the connector
+    connect.configureConnector(CONNECTOR_NAME, props);
+
+    // wait for tasks to spin up
+    waitForConnectorToStart(CONNECTOR_NAME, TASKS_MAX);
+
+    writeRecords(NUM_RECORDS);
+    waitForConnectorToFail(CONNECTOR_NAME, 1, trace);
   }
 
   protected void setDataStream() {
@@ -203,7 +216,7 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   }
 
   protected void writeRecordsFromStartIndex(int start, int numRecords) {
-    for (int i  = start; i < start + numRecords; i++) {
+    for (int i = start; i < start + numRecords; i++) {
       connect.kafka().produce(
           TOPIC,
           String.valueOf(i),
@@ -242,8 +255,8 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   }
 
   private static User getMinimalPrivilegesUser(boolean forDataStream) {
-        return new User(forDataStream ? ELASTIC_DATA_STREAM_MINIMAL_PRIVILEGES_NAME : ELASTIC_MINIMAL_PRIVILEGES_NAME,
-            Collections.singletonList(forDataStream ? ES_SINK_CONNECTOR_DS_ROLE : ES_SINK_CONNECTOR_ROLE));
+    return new User(forDataStream ? ELASTIC_DATA_STREAM_MINIMAL_PRIVILEGES_NAME : ELASTIC_MINIMAL_PRIVILEGES_NAME,
+        Collections.singletonList(forDataStream ? ES_SINK_CONNECTOR_DS_ROLE : ES_SINK_CONNECTOR_ROLE));
   }
 
   private static String getMinimalPrivilegesPassword(boolean forDataStream) {
